ISSUE #4827 - fixed minor bugs in panzoom
